// desktop/renderer/_Common/enum/status-api-code.enum.ts
export enum StatusAPICode {
  code = "code",
  hashing_password = 100,
  sign_in_request = 101,
  GET_EMPLOYEE_DETAILS = 102,
  UPDATE_APPLICABLE_SUBSIDY = 103,
  GET_CHECK_EMPLOYEE_ID = 104,
  GET_DEPARTMENT_LISTS = 105,
  CREATE_EMPLOYEE = 106,
  UPDATE_EMPLOYEE = 107,
  CREATE_SUBSIDY_TRANSACTION = 108,
  GET_EMPLOYEE_CATEGORY_LISTS = 109,
  UPLOAD_EXCEL_EMPLOYEE_CREATE = 110,
  GET_COST_CENTER_LISTS = 111,
  SUBSIDY_TRANSACTION_PAGINATION = 112,
  SUBSIDY_CHART_REPORT = 113,
  SUBSIDY_REPORT_DOWNLOAD = 114,
  CREATE_TRIGGER_SUBSIDY_CREDIT = 115,
  SUBSIDY_TYPE_PAGINATION = 116,
  UPDATE_SUBSIDY_TYPE = 117,
  UPDATE_USER_ACTIVE_STATUS = 118,
  UPDATE_SUBSIDY_CREDIT_REAL_TIME = 118,
  CREATE_SUBMIT_SUBSIDY_TRANSACTION_AUTH = 119,
  GET_CHECK_EMPLOYEE_ID_AUTH = 120,
  CREATE_TRIGGER_SUBSIDY_CREDIT_CRON = 121,
  GET_MENU_LISTS_PAGINATION = 122,
  UPLOAD_ID_IMAGE_EDIT_CROSSING = 123,
  CALLBACK_UPLOAD_FILES = 124,
  CREATE_MENU_CATEGORY = 125,
  GET_MENU_CATEGORY = 126,
  CREATE_MENU = 127,

}